Extend existing Audience and Level taxonomies
Remove new taxonomies added to 2024 theme
Remove helper functions mainly used for includes
