+++ wp-content/plugins/wporg-learn/inc/taxonomy.php
@@ -62,7 +62,7 @@ function register_lesson_audience() {
 		'items_list_navigation'      => __( 'Audiences list navigation', 'wporg-learn' ),
 	);
 
-	$args   = array(
+	$args = array(
 		'labels'            => $labels,
 		'hierarchical'      => true,
 		'public'            => true,
@@ -77,7 +77,7 @@ function register_lesson_audience() {
 		),
 	);
 
-	register_taxonomy( 'audience', array( 'lesson-plan' ), $args );
+	register_taxonomy( 'audience', array( 'lesson-plan', 'lesson', 'course' ), $args );
 }
 
 /**
@@ -241,7 +241,7 @@ function register_lesson_level() {
 		'items_list_navigation'      => __( 'Experience Levels list navigation', 'wporg-learn' ),
 	);
 
-	$args   = array(
+	$args = array(
 		'labels'            => $labels,
 		'hierarchical'      => true,
 		'public'            => true,
@@ -256,7 +256,7 @@ function register_lesson_level() {
 		),
 	);
 
-	register_taxonomy( 'level', array( 'lesson-plan' ), $args );
+	register_taxonomy( 'level', array( 'lesson-plan', 'lesson', 'course' ), $args );
 }
 
 /**
@@ -648,4 +648,52 @@ function tax_save_term_fields( $term_id ) {
 		'sticky',
 		rest_sanitize_boolean( $is_sticky )
 	);
+}
+
+/**
+ * Get available taxonomy terms for a post type.
+ *
+ * @param string $taxonomy The taxonomy.
+ * @param string $post_type The post type.
+ * @param string $post_status The post status.
+ * @return array The available taxonomy terms.
+ */
+function get_available_taxonomy_terms( $taxonomy, $post_type, $post_status = null ) {
+	$posts = get_posts( array(
+		'post_status'    => $post_status ?? 'any',
+		'post_type'      => $post_type,
+		'posts_per_page' => -1,
+	) );
+
+	if ( empty( $posts ) ) {
+		return array();
+	}
+
+	$term_ids = array();
+	foreach ( $posts as $post ) {
+		$post_terms = wp_get_post_terms( $post->ID, $taxonomy, array( 'fields' => 'ids' ) );
+
+		if ( ! is_wp_error( $post_terms ) ) {
+			$term_ids = array_merge( $term_ids, $post_terms );
+		}
+	}
+
+	if ( empty( $term_ids ) ) {
+		return array();
+	}
+
+	$term_ids = array_unique( $term_ids );
+
+	$terms = get_terms( array(
+		'taxonomy'   => $taxonomy,
+		'include'    => $term_ids,
+		'hide_empty' => false,
+	) );
+
+	$levels = array();
+	foreach ( $terms as $term ) {
+		$levels[ $term->slug ] = $term->name;
+	}
+
+	return $levels;
 }
